refactor(overview-ui): refactoring home page to pages folder
Following a comment from @gruvw, I have refactored the home page to the pages folder
and moved the bottom_navigation_bar widget to the views folder.
