+++ lib/views/pages/home/home_page.dart
@@ -1,8 +1,8 @@
 import "package:flutter/material.dart";
 import "package:hooks_riverpod/hooks_riverpod.dart";
-import "package:proxima/views/home/bottom_navigation_bar/navigation_bottom_bar.dart";
-import "package:proxima/views/home/posts/home_feed.dart";
-import "package:proxima/views/home/top_bar/home_top_bar.dart";
+import "package:proxima/views/bottom_navigation_bar/navigation_bottom_bar.dart";
+import "package:proxima/views/pages/home/posts/home_feed.dart";
+import "package:proxima/views/pages/home/top_bar/home_top_bar.dart";
 
 class HomePage extends HookConsumerWidget {
   const HomePage({super.key});
