+++ test/component/login_test.dart
@@ -2,8 +2,8 @@ import "package:firebase_core/firebase_core.dart";
 import "package:flutter_test/flutter_test.dart";
 import "package:hooks_riverpod/hooks_riverpod.dart";
 import "package:proxima/main.dart";
-import "package:proxima/views/home/top_bar/home_top_bar.dart";
-import "package:proxima/views/pages/home_page.dart";
+import "package:proxima/views/pages/home/top_bar/home_top_bar.dart";
+import "package:proxima/views/pages/home/home_page.dart";
 import "package:proxima/views/pages/login_page.dart";
 
 import "utils/firebase/setup_firebase_mocks.dart";
